requirements.txt
fastapi
uvicorn
python-multipart
openai
python-pptx
python-docx
pymupdf